requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# API Integration
openai==1.3.5
requests==2.31.0

# Optional: Development Dependencies
# Run: pip install -r requirements.txt[dev]
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
